Refactor: move gtfs raw data filtering to separate class. fixes also tests
